fix(test): тест сброса статусов перенёс в handlers пакет (#46)
В services-пакете тест коннектился по сырому DATABASE_URL (в CI это несуществующая
auto_registry) и упал. Перенёс в handlers_test через testutil.SetupTestApp/CleanDB -
правильная auto_registry_test и сериальная изоляция (CI гоняет пакеты параллельно -p4,
CleanDB соседнего пакета снёс бы строки).
